renamed usecase to step
