+++ src/Tests/Step/MakeAPRTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace PUGX\Bot\Tests\UseCase;
+namespace PUGX\Bot\Tests\Step;
 
 use PUGX\Bot\Package;
 use PUGX\Bot\LocalPackage;
-use PUGX\Bot\UseCase\MakeAPR;
+use PUGX\Bot\Step\MakeAPR;
 
 class MakeAPRTest extends BaseTestCase
 {
